requirements.txt
certifi==2021.5.30
cycler==0.10.0
et-xmlfile==1.1.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1630173724167/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1628881758439/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626271780153/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.3.56
openpyxl==3.0.7
pandas==1.3.2
patsy==0.5.1
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1630696607296/work
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1628206382406/work
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1629095986539/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1630699915890/work
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094706440/work
typing-extensions==3.10.0.2
yapf @ file:///tmp/build/80754af9/yapf_1615749224965/work
